+++ timber/lib/src/log_tree.dart
@@ -3,6 +3,9 @@ import 'package:logging/logging.dart' as $logging;
 
 import 'tree.dart';
 
-mixin LogTree {
+abstract class LogTree {
   void performLog($logging.LogRecord record);
+
+  /// Internal Logger Name
+  String get loggerName;
 }