+++ timber/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.3.1
+
+- fix: infinite loop perform log
+
 ## 0.3.0+1
 
 - fix: type check