+++ timber_crashlytics/CHANGELOG.md
@@ -1,3 +1,7 @@
-## [0.0.1] - TODO: Add release date.
+## [0.0.2]
 
-* TODO: Describe initial release.
+- fix: infinite loop perform log
+
+## [0.0.1] 
+
+